backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==0.8.11
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
joblib==1.3.2
aiofiles==23.2.1
nltk==3.8.1
xgboost==2.0.3
imbalanced-learn==0.11.0